Add C Split and oven_cook_split()
For now, it's a lot of Python C API boilerplate for not a lot of
converted python code, but I need a first step. It's rather hard to
start re-writing without pulling Account, Transation and Entry all at
once. If I do that, it will be a monster commit and it will fail because
it's too much complexity at once.