Make Entry.index readonly
This attribute serves only in a very narrow case: tie breaker for two entries
being generated for the same account and same transaction. Because tww
such entries are always going to be generated in the same cook_splits()
call (same txn), we can set this during Entry generation instead of
setting it in EntryList.add_entry(), thus making the attr readonly.