Account.group -> Account.groupname
Makes transition to ccore easier by not making group rewrite a
prerequisite to deep C-ification.

This can be easily done because account grouping is such a low impact
feature. The only consequence, besides lookup time, is that we need to
make sure that groupnames are updated after a group rename.