Introduce Entry type in ccore
Now, the gate is really opened to wash out a lot of py_ccore.c into
proper C units. Entry refers Split and Transaction directly, so lots of
logic can be de-python-ized.

Those 3.5k lines in py_ccore.c are becoming heavy. I expect this figure
to be the peak. Now, it's going to shrink.