Convert Transaction.change() to C
Whew, that was a big one. I thought I had some leeway in some of the
assignment logic w.r.t. amount/from/to change, but nope: any little
change in the logic make tests fail somewhere. This results in a rather
complicated C code to re-implement this.

Eventually, I'll have to revisit the reasons for the amount/from/to
change logic to work this way, but now is not the time.