Return a new instance in PyEntry_transaction()
Instead of returning `heldtxn`, return a new `PyTransaction` instance.
With the combination of not deallocating transactions in
`PyTransaction_dealloc()`, all tests pass fine.

This change doesn't do much by itself, but it proves that the app works
fine with volatile `PyTransaction` instances. Thus, it sets the roadmap
to achieve the goal of getting rid of `PyTransaction` references: either
find a way to stop relying on `recurrence` or convert `Recurrence` to C.
After that, we'll be able to greatly consolidate the "pure C" part of
moneyguru around the `Transaction` struct.