Push PyAccountList.entries down to pure C
This introduces glib to the project, which we use as a hash map
implementation (we used Python dictionary before that).

I hesitated between Glib's hash table, glibc `hcreate` and `uthash`. To
have more than one hash map, glibc requires us to enable `_GNU_SOURCE`
and I feel like it's a bad idea. Something like `uthash` would fit the
bill, but as it happens, many utilities in Glib will be useful in
moneyguru, so I opted for this.