Make Broadcaster notify its listeners in a predictable order
Having `listeners` as a set is a big problem for my "de-notification"
refactoring because I can't "flatten" the notification propagation (for
example, by making `MainWindow` a simple `Listener` and make its
children listen to `Document` directly) without wreaking havoc: children
start reacting to notifications before the main window and the code
isn't written for that).

So, to continue that flattening, I'm making `listeners` into a list.