Normalize refresh calls on txn table notifications
We used to have a `show_selected_row()` call on some refresh situation,
but not in others. I don't see why it was like this, I don't see why we
wouldn't want to call `show_selected_row()` after a txn table refresh,
so we now call it all the time. It simplifies the code.