Change the behavior of Document.stop_edition()
Since the first release of moneyGuru, the behavior has been the same: if
a txn, an entry or an account is being edited inline when we invoke a
"disruptive action" (an action that forces us to end the inline editing
operation, just as edit_item(), undo(), etc.), we save the content bein
edited before that action.

We're changing this to "cancel the edit operation" for two reasons.

First, it seems more intuitive to cancel an edit that hasn't been
explicitly committed by the user. It seems to me that this is what the
user expects.

Second, and more importantly, this allows us to simplify our process. To
be able to commit edits, we need to do it *before* the disruptive
action. This is complicated to do without bringing in some kind of
callback (notification) system.

When we cancel edits on disruptive action, we don't need to do it before
the action any more. It can very well happen just before we tell the UI
to refresh itself as a result of that disruptive action. Much simpler.

Note that the difference is, in fact, theoretical. As of now, the only
active UI for moneyGuru is Qt and the way the UI is done is that the
active edit is committed as soon as we go out of focus, so the situation
where a disruptive action is called while an edit is active never
actually happens. That's a good reason not to complicate core code to
save edits for this theoretical situation. Well, anyways, *if* we had a
UI that allowed disruptive actions to take place while edits are active,
*then* this commit would change moneyGuru's behavior.