Remove Group.expanded
This is now stored in account_sheet_view as a non-model attribute.

This is the first step in the upcoming "de-modelization" of account
groups. To simplify the memory model, account groups will stop being
first-class models. They'll only exist as attributes of accounts,
through `groupname`.

They will not be part of the undo/redo chain any more.

Of course, we still want to be able to create empty groups (otherwise,
how could we assign accounts to groups?), but these empty group will be
stored in the GUI, as ephemeral entity that will never reach the
`Document`. If we create an empty group and close moneyguru, that group
will disappear.

This should result in an important simplification of the code. Groups
were over-engineered.